selenium_elpais/requirements.txt
# Selenium WebDriver
selenium==4.15.2

# BrowserStack SDK
browserstack-sdk==1.7.0
browserstack-local>=1.2.3

# Web Driver Manager (for local testing)
webdriver-manager==4.0.1

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3

# Progress Bar
tqdm==4.66.1

# Beautiful Soup (backup parser)
beautifulsoup4==4.12.2
lxml==4.9.3